Change algoriphm for aborted processes.

+++ services/src/main/java/ru/taskurotta/service/gc/AbstractGCTask.java
@@ -2,6 +2,7 @@ package ru.taskurotta.service.gc;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import ru.taskurotta.service.console.model.Process;
 import ru.taskurotta.service.dependency.links.Graph;
 import ru.taskurotta.service.dependency.links.GraphDao;
 import ru.taskurotta.service.storage.ProcessService;
@@ -36,25 +37,30 @@ public abstract class AbstractGCTask implements Runnable {
         }
         logger.trace("Start garbage collector for process [{}]", processId);
 
-        Graph graph = graphDao.getGraph(processId);
-        if (graph == null) {
-            logger.warn("Not found graph for process [{}], stop garbage collector for this process", processId);
-            if (processService.getStartTask(processId) == null) {
-                logger.warn("And processService has no start task for it [{}]", processId);
+        Process process = processService.getProcess(processId);
+        boolean isAborted = process.getState() == Process.ABORTED;
+
+        if (!isAborted) {
+            Graph graph = graphDao.getGraph(processId);
+            if (graph == null) {
+                logger.warn("Not found graph for process [{}], stop garbage collector for this process", processId);
+                if (processService.getStartTask(processId) == null) {
+                    logger.warn("And processService has no start task for it [{}]", processId);
+                }
+                return;
             }
-            return;
+
+            if (!graph.isFinished()) {
+                logger.error("Graph for process [{}] isn't finished, stop garbage collector for this process", processId);
+                return;
+            }
+
+            Set<UUID> finishedItems = graph.getFinishedItems();
+            taskDao.deleteDecisions(finishedItems, processId);
+            taskDao.deleteTasks(finishedItems, processId);
+
+            graphDao.deleteGraph(processId);
         }
-
-        if (!graph.isFinished()) {
-            logger.error("Graph for process [{}] isn't finished, stop garbage collector for this process", processId);
-            return;
-        }
-
-        Set<UUID> finishedItems = graph.getFinishedItems();
-        taskDao.deleteDecisions(finishedItems, processId);
-        taskDao.deleteTasks(finishedItems, processId);
-
-        graphDao.deleteGraph(processId);
 
         processService.deleteProcess(processId);
